feat: home에서만 쓰이는 RankBoardWithTier 컴포넌트 생성 (#8)
